Edited tg_id from integer to BigInteger in DB and fix pagination

+++ bot/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from bot.core.db import Base
@@ -8,6 +8,6 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    tg_id = Column(Integer, unique=True, index=True)
+    tg_id = Column(BigInteger, unique=True, index=True)
     first_name = Column(String(128))
     reminders = relationship("Reminder", back_populates="user",)
